+++ test/src/helpers.cc
@@ -33,6 +33,9 @@
 #include "helpers.h"
 #include "catch.hpp"
 
+namespace tiledb {
+namespace test {
+
 template <class T>
 void check_partitions(
     tiledb::sm::SubarrayPartitioner& partitioner,
@@ -449,7 +452,7 @@ void write_array(
     tiledb_ctx_t* ctx,
     const std::string& array_name,
     tiledb_layout_t layout,
-    const AttrBuffers& attr_buffers) {
+    const QueryBuffers& buffers) {
   // Open array
   tiledb_array_t* array;
   int rc = tiledb_array_alloc(ctx, array_name.c_str(), &array);
@@ -465,7 +468,7 @@ void write_array(
   CHECK(rc == TILEDB_OK);
 
   // Set buffers
-  for (const auto& b : attr_buffers) {
+  for (const auto& b : buffers) {
     if (b.second.var_ == nullptr) {  // Fixed-sized
       rc = tiledb_query_set_buffer(
           ctx,
@@ -510,7 +513,7 @@ void read_array(
     tiledb_array_t* array,
     const SubarrayRanges<T>& ranges,
     tiledb_layout_t layout,
-    const AttrBuffers& attr_buffers) {
+    const QueryBuffers& buffers) {
   // Create query
   tiledb_query_t* query;
   int rc = tiledb_query_alloc(ctx, array, TILEDB_READ, &query);
@@ -529,7 +532,7 @@ void read_array(
   }
 
   // Set buffers
-  for (const auto& b : attr_buffers) {
+  for (const auto& b : buffers) {
     if (b.second.var_ == nullptr) {  // Fixed-sized
       rc = tiledb_query_set_buffer(
           ctx,
@@ -566,7 +569,6 @@ void read_array(
 }
 
 // Explicit template instantiations
-
 template void check_subarray<int8_t>(
     tiledb::sm::Subarray& subarray, const SubarrayRanges<int8_t>& ranges);
 template void check_subarray<uint8_t>(
@@ -685,58 +687,62 @@ template void read_array<int8_t>(
     tiledb_array_t* array,
     const SubarrayRanges<int8_t>& ranges,
     tiledb_layout_t layout,
-    const AttrBuffers& attr_buffers);
+    const QueryBuffers& buffers);
 template void read_array<uint8_t>(
     tiledb_ctx_t* ctx,
     tiledb_array_t* array,
     const SubarrayRanges<uint8_t>& ranges,
     tiledb_layout_t layout,
-    const AttrBuffers& attr_buffers);
+    const QueryBuffers& buffers);
 template void read_array<int16_t>(
     tiledb_ctx_t* ctx,
     tiledb_array_t* array,
     const SubarrayRanges<int16_t>& ranges,
     tiledb_layout_t layout,
-    const AttrBuffers& attr_buffers);
+    const QueryBuffers& buffers);
 template void read_array<uint16_t>(
     tiledb_ctx_t* ctx,
     tiledb_array_t* array,
     const SubarrayRanges<uint16_t>& ranges,
     tiledb_layout_t layout,
-    const AttrBuffers& attr_buffers);
+    const QueryBuffers& buffers);
 template void read_array<int32_t>(
     tiledb_ctx_t* ctx,
     tiledb_array_t* array,
     const SubarrayRanges<int32_t>& ranges,
     tiledb_layout_t layout,
-    const AttrBuffers& attr_buffers);
+    const QueryBuffers& buffers);
 template void read_array<uint32_t>(
     tiledb_ctx_t* ctx,
     tiledb_array_t* array,
     const SubarrayRanges<uint32_t>& ranges,
     tiledb_layout_t layout,
-    const AttrBuffers& attr_buffers);
+    const QueryBuffers& buffers);
 template void read_array<int64_t>(
     tiledb_ctx_t* ctx,
     tiledb_array_t* array,
     const SubarrayRanges<int64_t>& ranges,
     tiledb_layout_t layout,
-    const AttrBuffers& attr_buffers);
+    const QueryBuffers& buffers);
 template void read_array<uint64_t>(
     tiledb_ctx_t* ctx,
     tiledb_array_t* array,
     const SubarrayRanges<uint64_t>& ranges,
     tiledb_layout_t layout,
-    const AttrBuffers& attr_buffers);
+    const QueryBuffers& buffers);
 template void read_array<float>(
     tiledb_ctx_t* ctx,
     tiledb_array_t* array,
     const SubarrayRanges<float>& ranges,
     tiledb_layout_t layout,
-    const AttrBuffers& attr_buffers);
+    const QueryBuffers& buffers);
 template void read_array<double>(
     tiledb_ctx_t* ctx,
     tiledb_array_t* array,
     const SubarrayRanges<double>& ranges,
     tiledb_layout_t layout,
-    const AttrBuffers& attr_buffers);
+    const QueryBuffers& buffers);
+
+}  // End of namespace test
+
+}  // End of namespace tiledb